requirements.txt
fastapi==0.95.2
uvicorn==0.27.1
python-dotenv==1.0.0
google-generativeai==0.3.2
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
numpy==1.24.3
Jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.5.2 